requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
ldap3==2.4.1
MarkupSafe==1.0
Pillow==5.1.0
pyasn1==0.4.2
Werkzeug==0.14.1
